fix(docs): Review transport API docs and update Express middleware setup
