+++ docs/api/koa.md
@@ -25,7 +25,6 @@ The `@feathersjs/koa` module contains the [KoaJS](https://koajs.com/) framework 
 import { feathers } from '@feathersjs/feathers'
 import { koa, errorHandler, bodyParser, rest } from '@feathersjs/koa'
 
-// Create an app that is a Feathers AND Express application
 const app = koa(feathers())
 
 app.use(errorHandler())
@@ -34,12 +33,7 @@ app.use(bodyParser())
 app.configure(rest())
 ```
 
-`@feathersjs/koa` also exposes the following middleware:
-
-- `rest` - A middleware to expose services as REST APIs
-- `authentication` - A middleware for parsing HTTP headers for Feathers authentication information
-- `bodyParser` - The [koa-bodyparser](https://github.com/koajs/bodyparser) middleware for parsing request bodies
-- `errorHandler` - A JSON error handler middleware. Should always be registered as the very first middleware.
+Also see the [additional middleware](#middleware) that `@feathersjs/koa` exposes.
 
 ## koa(app, koaApp)
 
@@ -52,57 +46,6 @@ If no Feathers application is passed, `koa() -> app` returns a plain Koa applica
 ## app.use(location|mw[, service])
 
 `app.use(location|mw[, service]) -> app` registers either a [service object](./services.md), or a Koa middleware. If a path and [service object](./services.md) is passed it will use Feathers registration mechanism, for a middleware function Koa.
-
-### Koa middleware
-
-In a Koa middleware, `ctx.feathers` is an object which will be extended as `params` in a service method call. 
-
-```ts
-import type { NextFunction } from '@feathersjs/koa'
-import type { Id, Params } from '@feathersjs/feathers'
-
-class TodoService {
-  async get(id: Id, params: Params & { fromMiddleware: string }) {
-    const { fromMiddleware } = params
-
-    return { id, fromMiddleware }
-  }
-}
-
-// Register Koa middleware
-app.use(async (ctx: any, next: NextFunction) => {
-  ctx.feathers.fromMiddleware = 'Hello from Koa middleware'
-
-  await next()
-})
-
-// Register a service
-app.use('todos', new TodoService())
-```
-
-### Service middleware
-
-When registering a service, it is also possible to pass custom Koa middleware that should run `before` the specific service method in the `koa` [service option](./application.md#usepath-service--options):
-
-
-```ts
-app.use('/todos', new TodoService(), {
-  koa: {
-    before: [async (ctx, next) => {
-      ctx.feathers // data that will be merged into sevice `params`
-      
-      // This will run all subsequent middleware and the service call
-      await next()
-
-      // Then we have additional properties available on the context
-      ctx.hook // the hook context from the method call
-      ctx.body // the return value
-    }]
-  }
-})
-```
-
-Note that the order of middleware will be `[...before, serviceMethod]`.
 
 ## app.listen(port)
 
@@ -117,6 +60,113 @@ const server = await app.listen(3030)
 
 `app.setup(server) -> app` is usually called internally by `app.listen` but in the cases described below needs to be called explicitly.
 
+## params
+
+In a Koa middleware, `ctx.feathers` is an object which will be extended as `params` in a service method call.
+
+```ts
+import { rest } from '@feathersjs/koa'
+import type { NextFunction } from '@feathersjs/koa'
+import type { Id, Params } from '@feathersjs/feathers'
+
+class TodoService {
+  async get(id: Id, params: Params & { fromMiddleware: string }) {
+    const { fromMiddleware } = params
+
+    return { id, fromMiddleware }
+  }
+}
+
+// Register Koa middleware
+app.use(async (ctx: any, next: NextFunction) => {
+  ctx.feathers = {
+    ...ctx.feathers,
+    fromMiddleware: 'Hello from Koa middleware'
+  }
+
+  await next()
+})
+app.configure(rest())
+
+// Register a service
+app.use('todos', new TodoService())
+```
+
+<BlockQuote type="warning" label="Important">
+
+Note that `app.configure(rest())` has to happen **after** any custom middleware.
+
+</BlockQuote>
+
+### params.query
+
+`params.query` will contain the URL query parameters sent from the client parsed using [koa-qs](https://github.com/koajs/qs).
+
+<BlockQuote type="warning" label="important">
+
+Only `params.query` is passed between the server and the client, other parts of `params` are not. This is for security reasons so that a client can't set things like `params.user` or the database options. You can always map from `params.query` to other `params` properties in a [hook](./hooks.md).
+
+</BlockQuote>
+
+### params.provider
+
+For any [service method call](./services.md) made through REST `params.provider` will be set to `rest`.
+
+### params.route
+
+Route placeholders in a service URL will be added to the services `params.route`. See the [FAQ entry on nested routes](../help/faq.md#how-do-i-do-nested-or-custom-routes) for more details on when and when not to use nested routes.
+
+```ts
+import { feathers } from '@feathersjs/feathers'
+import { koa, errorHandler, bodyParser, rest } from '@feathersjs/koa'
+
+const app = koa(feathers())
+
+app.use('users/:userId/messages', {
+  async get(id, params) {
+    console.log(params.query) // -> ?query
+    console.log(params.provider) // -> 'rest'
+    console.log(params.fromMiddleware) // -> 'Hello world'
+    console.log(params.route) // will be `{ userId: '1' }` for GET /users/1/messages
+
+    return {
+      id,
+      params,
+      read: false,
+      text: `Feathers is great!`,
+      createdAt: new Date().getTime()
+    }
+  }
+})
+
+app.listen(3030)
+```
+
+## Service middleware
+
+When registering a service, it is also possible to pass custom Koa middleware that should run `before` the specific service method in the `koa` [service option](./application.md#usepath-service--options):
+
+```ts
+app.use('/todos', new TodoService(), {
+  koa: {
+    before: [
+      async (ctx, next) => {
+        ctx.feathers // data that will be merged into sevice `params`
+
+        // This will run all subsequent middleware and the service call
+        await next()
+
+        // Then we have additional properties available on the context
+        ctx.hook // the hook context from the method call
+        ctx.body // the return value
+      }
+    ]
+  }
+})
+```
+
+Note that the order of middleware will be `[...before, serviceMethod]`.
+
 ## Middleware
 
 ### rest
@@ -130,14 +180,14 @@ app.configure(rest())
 Configures the middleware for handling service calls via HTTP. It will also register authentication header parsing. The following (optional) options are available:
 
 - `formatter` - A middleware that formats the response body
-- `authentication` - The authentication `service` and `strategies` to use for parsing authentication information 
+- `authentication` - The authentication `service` and `strategies` to use for parsing authentication information
 
 ### errorHandler
 
 ```ts
 import { errorHandler } from '@feathersjs/koa'
 
-app.configure(errorHandler())
+app.use(errorHandler())
 ```
 
 A middleware that formats errors as a Feathers error and sets the proper status code. Needs to be the first middleware registered in order to catch all other errors.
@@ -149,3 +199,11 @@ A middleware that allows to authenticate a user (or other authentication entity)
 ### bodyParser
 
 A reference to the [koa-bodyparser](https://github.com/koajs/bodyparser) module.
+
+### cors
+
+A reference to the [@koa/cors](https://github.com/koajs/cors) module.
+
+### serveStatic
+
+A reference to the [koa-static](https://github.com/koajs/static) module.
