+++ docs/api/socketio.md
@@ -27,9 +27,9 @@ This page describes how to set up a Socket.io server. The [Socket.io client chap
 
 `@feathersjs/socketio` can be used standalone or together with a Feathers framework integration like [Express](./express.md).
 
-### app.configure(socketio())
+### socketio()
 
-Sets up the Socket.io transport with the default configuration using either the server provided by [app.listen](./application.md#listenport) or passed in [app.setup(server)](./application.md#setupserver).
+`app.configure(socketio())` sets up the Socket.io transport with the default configuration using either the server provided by [app.listen](./application.md#listenport) or passed in [app.setup(server)](./application.md#setupserver).
 
 ```ts
 import { feathers } from '@feathersjs/feathers'
@@ -48,9 +48,9 @@ Once the server has been started with `app.listen()` or `app.setup(server)` the 
 
 </BlockQuote>
 
-### app.configure(socketio(callback))
+### socketio(callback)
 
-Sets up the Socket.io transport with the default configuration and call `callback` with the [Socket.io server object](http://socket.io/docs/server-api/).
+`app.configure(socketio(callback))` sets up the Socket.io transport with the default configuration and call `callback` with the [Socket.io server object](http://socket.io/docs/server-api/).
 
 ```ts
 import { feathers } from '@feathersjs/feathers'
@@ -58,54 +58,60 @@ import socketio from '@feathersjs/socketio'
 
 const app = feathers()
 
-app.configure(socketio(io => {
-  io.on('connection', socket => {
-    // Do something here
-  })
+app.configure(
+  socketio((io) => {
+    io.on('connection', (socket) => {
+      // Do something here
+    })
 
-  // Registering Socket.io middleware
-  io.use(function (socket, next) {
-    // Exposing a request property to services and hooks
-    socket.feathers.referrer = socket.request.referrer
-    next()
+    // Registering Socket.io middleware
+    io.use(function (socket, next) {
+      // Exposing a request property to services and hooks
+      socket.feathers.referrer = socket.request.referrer
+      next()
+    })
   })
-}))
+)
 
 app.listen(3030)
 ```
 
-<BlockQuote type="warning">
+<BlockQuote type="danger">
 
-We recommend to avoid listening and sending events on the `socket` directly since it circumvents Feathers secure dispatch mechanisms available through [channels](./channels.md) and [hooks](./hooks.md).
+Try to avoid listening and sending events on the `socket` directly since it circumvents Feathers secure dispatch mechanisms available through [channels](./channels.md) and [hooks](./hooks.md).
 
 </BlockQuote>
 
-### app.configure(socketio(options [, callback]))
+### socketio(options [, callback])
 
-Sets up the Socket.io transport with the given [Socket.io options object](https://github.com/socketio/engine.io#methods-1) and optionally calls the callback described above.
+`app.configure(socketio(options [, callback]))` sets up the Socket.io transport with the given [Socket.io options object](https://github.com/socketio/engine.io#methods-1) and optionally calls the callback described above.
 
 This can be used to e.g. configure the path where Socket.io is initialize (`socket.io/` by default). The following changes the path to `ws/`:
-
 
 ```ts
 import { feathers } from '@feathersjs/feathers'
 import socketio from '@feathersjs/socketio'
 
-const app = feathers();
+const app = feathers()
 
-app.configure(socketio({
-  path: '/ws/'
-}, io => {
-  // Do something here
-  // This function is optional
-}))
+app.configure(
+  socketio(
+    {
+      path: '/ws/'
+    },
+    (io) => {
+      // Do something here
+      // This function is optional
+    }
+  )
+)
 
 app.listen(3030)
 ```
 
-### app.configure(socketio(port, [options], [callback]))
+### socketio(port, [options], [callback])
 
-Creates a new Socket.io server on a separate port. Options and a callback are optional and work as described above.
+`app.configure(socketio(port, [options], [callback]))` creates a new Socket.io server on a separate port. Options and a callback are optional and work as described above.
 
 ```ts
 import { feathers } from '@feathersjs/feathers'
@@ -122,12 +128,14 @@ app.listen(3030)
 [Socket.io middleware](https://socket.io/docs/v4/middlewares/) can modify the `feathers` property on the `socket` which will then be used as the service call `params`:
 
 ```ts
-app.configure(socketio(io => {
-  io.use((socket, next) => {
-    socket.feathers.user = { name: 'David' }
-    next()
+app.configure(
+  socketio((io) => {
+    io.use((socket, next) => {
+      socket.feathers.user = { name: 'David' }
+      next()
+    })
   })
-}))
+)
 
 app.use('messages', {
   async create(data, params, callback) {
@@ -154,12 +162,12 @@ app.service('users').hooks({
   before: {
     remove(context) {
       // check for if(context.params.provider) to prevent any external call
-      if(context.params.provider === 'socketio') {
-        throw new Error('You can not delete a user via Socket.io');
+      if (context.params.provider === 'socketio') {
+        throw new Error('You can not delete a user via Socket.io')
       }
     }
   }
-});
+})
 ```
 
 ### params.query
@@ -175,3 +183,15 @@ Only `params.query` is passed between the server and the client, other parts of 
 ### params.connection
 
 `params.connection` is the connection object that can be used with [channels](./channels.md). It is the same object as `socket.feathers` in a Socket.io middleware as [shown in the `params` section](#params).
+
+### params.headers
+
+`params.headers` contains the headers from the original handshake. This is usually sent with the `extraHeaders` option when initialising the connection on the client:
+
+```ts
+const socket = io('http://localhost:9777', {
+  extraHeaders: {
+    MyHeader: 'somevalue'
+  }
+})
+```
