+++ packages/express/src/declarations.ts
@@ -49,7 +49,7 @@ declare module '@feathersjs/feathers/lib/declarations' {
 
 declare module 'express-serve-static-core' {
   interface Request {
-    feathers?: Partial<FeathersParams> & { [key: string]: any }
+    feathers: Partial<FeathersParams> & { [key: string]: any }
     lookup?: RouteLookup
   }
