+++ packages/socketio/test/events.ts
@@ -1,7 +1,7 @@
 import { strict as assert } from 'assert'
 import { io, Socket } from 'socket.io-client'
 import { verify } from '@feathersjs/tests'
-import { RealTimeConnection } from '@feathersjs/transport-commons/src/channels/channel/base'
+import { RealTimeConnection } from '@feathersjs/feathers'
 
 export default (name: string, options: any) => {
   const call = (method: string, ...args: any[]) => {
