+++ packages/transport-commons/src/channels/index.ts
@@ -1,7 +1,7 @@
-import { Application, FeathersService, getServiceOptions } from '@feathersjs/feathers'
+import { Application, FeathersService, RealTimeConnection, getServiceOptions } from '@feathersjs/feathers'
 import { createDebug } from '@feathersjs/commons'
 import { compact, flattenDeep, noop } from 'lodash'
-import { Channel, RealTimeConnection } from './channel/base'
+import { Channel } from './channel/base'
 import { CombinedChannel } from './channel/combined'
 import { channelMixin, publishMixin, keys, PublishMixin, Event, Publisher } from './mixins'
 import EventEmitter from 'events'
