+++ packages/transport-commons/src/index.ts
@@ -1,6 +1,7 @@
 import { socket } from './socket'
 import { routing } from './routing'
-import { channels, Channel, CombinedChannel, RealTimeConnection } from './channels'
+import { channels, Channel, CombinedChannel } from './channels'
+import { RealTimeConnection } from '@feathersjs/feathers'
 
 export * as http from './http'
 export { socket, routing, channels, Channel, CombinedChannel, RealTimeConnection }
