+++ packages/transport-commons/src/socket/utils.ts
@@ -1,9 +1,14 @@
-import { HookContext, Application, createContext, getServiceOptions } from '@feathersjs/feathers'
+import {
+  HookContext,
+  Application,
+  RealTimeConnection,
+  createContext,
+  getServiceOptions
+} from '@feathersjs/feathers'
 import { NotFound, MethodNotAllowed, BadRequest } from '@feathersjs/errors'
 import { createDebug } from '@feathersjs/commons'
 import isEqual from 'lodash/isEqual'
 import { CombinedChannel } from '../channels/channel/combined'
-import { RealTimeConnection } from '../channels/channel/base'
 
 const debug = createDebug('@feathersjs/transport-commons')
